Update DIR endpoint request tests
